requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
datetime-truncate==1.1.1
Django==3.2.16
django-admin-charts==1.2.4
django-bower==5.2.0
django-elasticsearch-dsl==7.2.2
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-utils==0.0.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
promise==2.3
PyJWT==1.7.1
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.4.0
Unidecode==1.3.6
urllib3==1.26.14
